Add 'run compression'
For each possible byte, try replacing all runs of >= 1 instance of it
with just a single instance.

The idea is that repeated characters are often considered equivalent
to a single character (particularly for whitespace, but this works
well for e.g. identifiers).

Doing this both deletes a *lot* of data in human readable formats
and also makes it much easier to find ngrams in the next stage.