Add "alphabet minimization"
This tries to reduce the alphabet size of the file, replacing
characters with other lower value characters in the file.

We do this quite late on as it's a potentially very slow operation
(up to 30k shrink attempts if we're really unlucky!) but later on
the test is faster and the alphabet size is smaller.