Save history of shrinks
This is an idea borrowed from Hypothesis: Every time we successfully
shrink something we note it for later. This is because 'slippage'
might occur where two things that look like the same bug turn out
not to be. If we save the entire history of shrinks this matters
less because we can examine the shrunk file and refine our test if
it was not the bug we were looking for.