Quite a significant update
The main point of this update is to provide a new pass (charwise)
and greatly simplify the ngram calculations because they can assume
they run on an already fairly minimized source.

It adds a bunch of functionality along the way. In particular you
can now run only the passes you want.