Just track whether we've seen a value
Caching the result of classify turns out to be quite
expensive with multishrinking, because the values are
potentially very large and usually distinct.

It's also unneeded. Because we track the best example
for each label, all we need to know is if we've ever
seen the string before. If we have, then it's not an
improvement and we can avoid it.